added the zipkin API and corresponding endpoints "getServices" and "getSpans" according to https://zipkin.io/zipkin-api/#/default/ Also added the helper function JsonUtil.writeJsonValues to parse json objects to json lists.
