+++ kaldb/src/main/java/com/slack/kaldb/server/Kaldb.java
@@ -15,7 +15,6 @@ import com.slack.kaldb.clusterManager.ReplicaDeletionService;
 import com.slack.kaldb.clusterManager.ReplicaEvictionService;
 import com.slack.kaldb.clusterManager.SnapshotDeletionService;
 import com.slack.kaldb.elasticsearchApi.ElasticsearchApiService;
-import com.slack.kaldb.elasticsearchApi.ZipkinService;
 import com.slack.kaldb.logstore.LogMessage;
 import com.slack.kaldb.logstore.search.KaldbDistributedQueryService;
 import com.slack.kaldb.logstore.search.KaldbLocalQueryService;
@@ -34,6 +33,7 @@ import com.slack.kaldb.proto.config.KaldbConfigs;
 import com.slack.kaldb.recovery.RecoveryService;
 import com.slack.kaldb.util.RuntimeHalterImpl;
 import io.micrometer.core.instrument.MeterRegistry;
+import com.slack.kaldb.zipkinApi.ZipkinService;
 import io.micrometer.core.instrument.Metrics;
 import io.micrometer.core.instrument.binder.jvm.ClassLoaderMetrics;
 import io.micrometer.core.instrument.binder.jvm.JvmGcMetrics;
